a lot of shit

+++ src/main/java/spigey/bot/Commands/ChangePasswordCommand.java
@@ -20,18 +20,17 @@ public class ChangePasswordCommand implements Command {
         String oldPassword = sys.decrypt(db.read("passwords", "password_" + username), env.ENCRYPTION_KEY);
         if(Objects.equals(username, "0")){event.reply("You need to be logged in to use this command!").setEphemeral(true).queue(); return 0;}
         if(!event.getOption("old-password").getAsString().equals(oldPassword)){event.reply("Invalid password provided.").setEphemeral(true).queue(); return 0;}
-        AtomicInteger i = new AtomicInteger();
+        if(event.getOption("new-password").getAsString().length() < 6 || event.getOption("new-password").getAsString().length() > 40){event.reply("Invalid password. Must be between 6 and 24 characters in length").setEphemeral(true).queue(); return 0;}
         util.userExec(username, user -> {
             if(!user.getId().equals(self)) {
-                i.getAndIncrement();
                 db.remove(user.getId(), "account");
             }
         });
         db.write("passwords", "password_" + username, sys.encrypt(event.getOption("new-password").getAsString(), env.ENCRYPTION_KEY));
-        event.reply("You have successfully changed your password to ||`" + event.getOption("new-password").getAsString() + "`|| and logged out `" + i.get() + "` users!").setEphemeral(true).queue();
+        event.reply("You have successfully changed your password to ||`" + event.getOption("new-password").getAsString() + "`||!").setEphemeral(true).queue();
         EmbedBuilder embed = new EmbedBuilder()
                 .setTitle("Password Changed") //                                                                      ↓ that password is encrypted
-                .setDescription(String.format("**Username**: `%s`\n**Old Password Strength**: `%s / 10`\n**New Password Strength**: `%s / 10`**Users Logged Out**: `%s`\n||%s|| -> ||%s||", username, sys.passStrength(oldPassword), sys.passStrength(event.getOption("new-password").getAsString()), i.get(), oldPassword, event.getOption("new-password").getAsString()))
+                .setDescription(String.format("**Username**: `%s`\n**Old Password Strength**: `%s%%`\n**New Password Strength**: `%s%%`\n||%s|| -> ||%s||", username, sys.passStrength(oldPassword), sys.passStrength(event.getOption("new-password").getAsString()), oldPassword, event.getOption("new-password").getAsString()))
                 .setColor(EmbedColor.GOLD);
         event.getJDA().getGuildById("1219338270773874729").getTextChannelById("1246196496592801834").sendMessage(event.getUser().getId()).addEmbeds(embed.build()).addActionRow(
                 Button.danger("snipe", "Snipe").withEmoji(Emoji.fromUnicode("U+1F52B")),
